separate stores into the files
