+++ src/components/DishList.vue
@@ -12,7 +12,7 @@
 
 <script setup>
 import { onMounted } from 'vue'
-import { useDishStore } from '@store'
+import { useDishStore } from '@store/dish'
 import { storeToRefs } from 'pinia'
 import DishCard from './DishCard.vue'
