+++ src/components/FilterOptions.vue
@@ -77,7 +77,7 @@
 import { ref, computed } from 'vue'
 import { storeToRefs } from 'pinia'
 import { deepUnref } from 'vue-deepunref'
-import { useDishStore } from '@store'
+import { useDishStore } from '@store/dish'
 
 const store = useDishStore()
 const { getDishes } = store
