+++ src/views/HomeView.vue
@@ -51,7 +51,7 @@ import { ref, watch } from 'vue'
 import { deepUnref } from 'vue-deepunref'
 import Filters from '@components/FilterOptions.vue'
 import DishList from '@components/DishList.vue'
-import { useDishStore } from '@store'
+import { useDishStore } from '@store/dish'
 import { history } from 'instantsearch.js/es/lib/routers'
 import { singleIndex } from 'instantsearch.js/es/lib/stateMappings'
 import router from '@src/router'
